// ref_truth/types.ts
export enum PlantCode {
    CIKOKOL = 'CIKOKOL',
    SEMARANG = 'SEMARANG',
    CIKUPA = 'CIKUPA',
    AGRO = 'AGRO'
}

export enum MachineStatus {
    RUNNING = 'RUNNING',
    IDLE = 'IDLE',
    BREAKDOWN = 'BREAKDOWN',
    OFFLINE = 'OFFLINE'
}

export enum AlarmSeverity {
    INFO = 'INFO',
    WARNING = 'WARNING',
    CRITICAL = 'CRITICAL'
}

export enum UserRole {
    ADMINISTRATOR = 'Administrator',
    SUPERVISOR = 'Supervisor',
    OPERATOR = 'Operator',
    MAINTENANCE = 'Maintenance',
    QC = 'Quality Control',
    MANAGEMENT = 'Management',
    VIEWER = 'Viewer'
}

export enum MachineType {
    EXTRUDER = 'EXTRUDER',
    FRYER = 'FRYER',
    SEASONING = 'SEASONING',
    PACKING = 'PACKING',
    GENERIC = 'GENERIC'
}

export interface User {
    username: string;
    name: string;
    role: UserRole;
    plantAccess?: PlantCode[]; // Array of PlantIDs this user can access
}